Add TrajservLayer mixin
