+++ src/mapbox/layers/TrajservLayer.js
@@ -1,22 +1,13 @@
-import qs from 'query-string';
 import { fromLonLat } from 'ol/proj';
-import { buffer, getWidth } from 'ol/extent';
 // import { LineString } from 'ol/geom';
+import { buffer, getWidth } from 'ol/extent';
 import TrackerLayer from './TrackerLayer';
-import { getDateString, getUTCTimeString } from '../../common/timeUtils';
-import {
-  getRadius,
-  getBgColor,
-  getDelayColor,
-  getDelayText,
-  getTextColor,
-  getTextSize,
-} from '../../common/trackerConfig';
-import { TrajservAPI } from '../../api';
+import mixin from '../../common/mixins/TrajservLayerMixin';
+import { getUTCTimeString } from '../../common/timeUtils';
 import { getSourceCoordinates, getResolution } from '../utils';
 
 /**
- * Responsible for loading tracker data from Realtime service.
+ * Responsible for loading and display data from a Trajserv service.
  *
  * @example
  * import { TrajservLayer } from 'mobility-toolbox-js/src/mapbox';
@@ -28,81 +19,13 @@ import { getSourceCoordinates, getResolution } from '../utils';
  *
  * @see https://mobility-toolbox-js.geops.de/api/class/src/api/trajserv/TrajservAPI.js~TrajservAPI.html
  * @see https://mobility-toolbox-js.geops.de/examples/mapbox-tracker
+ *
+ * @extends {TrackerLayer}
+ * @implements {TrajservLayerInterface}
  */
-class TrajservLayer extends TrackerLayer {
-  /**
-   * Create a filter based on train and operator
-   * @param {string} line
-   * @param {string} route
-   * @param {string} operator
-   * @param {string} regexLine
-   * @private
-   */
-  static createFilter(line, trip, operator, regexLine) {
-    const filterList = [];
-
-    if (!line && !trip && !operator && !regexLine) {
-      return null;
-    }
-
-    if (regexLine) {
-      const regexLineList =
-        typeof regexLine === 'string' ? [regexLine] : regexLine;
-      const lineFilter = (t) =>
-        regexLineList.some((tr) => new RegExp(tr, 'i').test(t.name));
-      filterList.push(lineFilter);
-    }
-
-    if (line) {
-      const lineFiltersList = typeof line === 'string' ? line.split(',') : line;
-      const lineList = lineFiltersList.map((l) =>
-        l.replace(/\s+/g, '').toUpperCase(),
-      );
-      const lineFilter = (l) =>
-        lineList.some((filter) => filter === l.name.toUpperCase());
-      filterList.push(lineFilter);
-    }
-
-    if (trip) {
-      const tripFilters = typeof trip === 'string' ? trip.split(',') : trip;
-      const tripList = tripFilters.map((rt) => parseInt(rt, 10));
-      const tripFilter = (t) => {
-        const tripId = parseInt(t.routeIdentifier.split('.')[0], 10);
-        return tripList.some((tr) => tr === tripId);
-      };
-      filterList.push(tripFilter);
-    }
-
-    if (operator) {
-      const operatorList = typeof operator === 'string' ? [operator] : operator;
-      const operatorFilter = (t) =>
-        operatorList.some((op) => new RegExp(op, 'i').test(t.operator));
-      filterList.push(operatorFilter);
-    }
-
-    return (t) => {
-      for (let i = 0; i < filterList.length; i += 1) {
-        if (!filterList[i](t)) {
-          return false;
-        }
-      }
-      return true;
-    };
-  }
-
+class TrajservLayer extends mixin(TrackerLayer) {
   constructor(options = {}) {
     super({ ...options });
-
-    this.options = options;
-    this.showVehicleTraj =
-      options.showVehicleTraj !== undefined ? options.showVehicleTraj : true;
-    this.delayDisplay = options.delayDisplay || 300000;
-    this.requestIntervalSeconds = 3;
-    this.useDelayStyle = options.useDelayStyle || false;
-    this.delayOutlineColor = options.delayOutlineColor || '#000000';
-    this.width = options.width;
-    this.height = options.height;
-    this.api = new TrajservAPI({ url: options.url, apiKey: options.apiKey });
     this.onMapClick = this.onMapClick.bind(this);
     this.onMove = this.onMove.bind(this);
     this.onMoveEnd = this.onMoveEnd.bind(this);
@@ -114,21 +37,11 @@ class TrajservLayer extends TrackerLayer {
    * @param {String} beforeLayerId .
    */
   init(map, beforeLayerId) {
-    super.init(map);
-    if (!this.map) {
+    if (!map) {
       return;
     }
 
-    // Sort the trajectories.
-    if (this.sortFc) {
-      this.sort = this.sortFc;
-    } else if (this.useDelayStyle) {
-      // Automatic sorting depending on delay, higher delay on top.
-      this.sort = (a, b) => {
-        if (a.delay === null) return 1;
-        return a.delay < b.delay ? 1 : -1;
-      };
-    }
+    super.init(map);
 
     const { width, height } = map.getCanvas();
     this.tracker.canvas.width = width;
@@ -156,45 +69,17 @@ class TrajservLayer extends TrackerLayer {
     );
   }
 
-  addTrackerFilters() {
-    // Setting filters from the permalink.
-    const parameters = qs.parse(window.location.search.toLowerCase());
-    const lineParam = parameters[TrajservLayer.LINE_FILTER];
-    const routeParam = parameters[TrajservLayer.ROUTE_FILTER];
-    const opParam = parameters[TrajservLayer.OPERATOR_FILTER];
-    const { regexPublishedLineName } = this.options;
-
-    if (lineParam || routeParam || opParam || regexPublishedLineName) {
-      this.filter = TrajservLayer.createFilter(
-        lineParam ? lineParam.split(',') : undefined,
-        routeParam ? routeParam.split(',') : undefined,
-        opParam ? opParam.split(',') : undefined,
-        regexPublishedLineName,
-      );
-    } else {
-      this.filter = null;
-    }
-  }
-
   start() {
     if (!this.map) {
       return;
     }
-
-    this.addTrackerFilters();
-
-    super.start(this.map);
-    this.startUpdateTrajectories();
-
+    super.start();
     this.map.on('click', this.onMapClick);
     this.map.on('move', this.onMove);
     this.map.on('moveend', this.onMoveEnd);
   }
 
   stop() {
-    this.journeyId = null;
-    this.stopUpdateTrajectories();
-    this.abortFetchTrajectories();
     if (this.map) {
       this.map.off('click', this.onClick);
       this.map.off('move', this.onMove);
@@ -248,6 +133,10 @@ class TrajservLayer extends TrackerLayer {
       this.journeyId = vehicle.journeyIdentifier;
       this.updateTrajectoryStations(this.selectedVehicleId).then(
         (vehicleWithStations) => {
+          this.stationsCoords = [];
+          vehicleWithStations.stations.forEach((station) => {
+            this.stationsCoords.push(fromLonLat(station.coordinates));
+          });
           this.clickCallbacks.forEach((callback) =>
             callback(vehicleWithStations, this, evt),
           );
@@ -257,6 +146,33 @@ class TrajservLayer extends TrackerLayer {
       this.selectedVehicleId = null;
       this.clickCallbacks.forEach((callback) => callback(null, this, evt));
     }
+  }
+
+  /**
+   * @override
+   */
+  getParams(extraParams = {}) {
+    const bounds = this.map.getBounds().toArray();
+    const southWest = fromLonLat(bounds[0]);
+    const northEast = fromLonLat(bounds[1]);
+    const ext = [...southWest, ...northEast];
+    const bbox = buffer(ext, getWidth(ext) / 10).join(',');
+    const zoom = this.map.getZoom();
+
+    return super.getParams({
+      ...extraParams,
+      bbox,
+      s: zoom < 10 ? 1 : 0,
+      z: zoom,
+    });
+  }
+
+  /**
+   * @override
+   */
+  defaultStyle(props) {
+    const zoom = this.map.getZoom();
+    return super.defaultStyle(props, zoom);
   }
 
   /**
@@ -324,31 +240,6 @@ class TrajservLayer extends TrackerLayer {
     // vectorSource.addFeature(lineFeat);
   }
 
-  abortFetchTrajectories() {
-    if (this.abortController) {
-      this.abortController.abort();
-    }
-  }
-
-  /**
-   * Fetch stations information with a trajectory ID
-   * @param {number} trajId The ID of the trajectory
-   * @private
-   */
-  updateTrajectoryStations(trajId) {
-    const params = this.getParams({
-      id: trajId,
-      time: getUTCTimeString(new Date()),
-    });
-    return this.api.fetchTrajectoryStations(params).then((trajStations) => {
-      this.stationsCoords = [];
-      trajStations.stations.forEach((station) => {
-        this.stationsCoords.push(fromLonLat(station.coordinates));
-      });
-      return trajStations;
-    });
-  }
-
   /**
    * Highlight the trajectory of journey.
    * @param {String} journeyId The id of the journey.
@@ -378,200 +269,6 @@ class TrajservLayer extends TrackerLayer {
         this.map.removeLayer('highlight-trajectory');
       });
   }
-
-  /**
-   * Returns the URL Parameters
-   * @param {Object} extraParams
-   * @returns {Object}
-   * @private
-   */
-  getParams(extraParams = {}) {
-    const bounds = this.map.getBounds().toArray();
-    const southWest = fromLonLat(bounds[0]);
-    const northEast = fromLonLat(bounds[1]);
-    const ext = [...southWest, ...northEast];
-    const bbox = buffer(ext, getWidth(ext) / 10).join(',');
-    const intervalMs = this.speed * 20000; // 20 seconds, arbitrary value, could be : (this.requestIntervalSeconds + 1) * 1000;
-    const now = this.currTime;
-
-    let diff = true;
-
-    if (
-      this.later &&
-      now.getTime() > this.later.getTime() - 3000 * this.speed
-    ) {
-      diff = false;
-    }
-    if (
-      !this.later ||
-      !diff ||
-      this.later.getTime() - now.getTime() > intervalMs
-    ) {
-      const later = new Date(now.getTime() + intervalMs);
-      this.later = later;
-    }
-
-    const params = {
-      ...extraParams,
-      bbox,
-      btime: getUTCTimeString(now),
-      etime: getUTCTimeString(this.later),
-      date: getDateString(now),
-      rid: 1,
-      a: 1,
-      cd: 1,
-      nm: 1,
-      fl: 1,
-      s: this.map.getZoom() < 10 ? 1 : 0,
-      z: this.map.getZoom(),
-      // toff: this.currTime.getTime() / 1000,
-    };
-
-    // Allow to load only differences between the last request,
-    // but currently the Tracker render method doesn't manage to render only diff.
-    /* if (diff) {
-      // Not working
-      params.diff = this.lastRequestTime;
-    } */
-    return params;
-  }
-
-  /**
-   * Start the update of trajectories.
-   * @private
-   */
-  startUpdateTrajectories() {
-    this.stopUpdateTrajectories();
-
-    this.updateTrajectories();
-    this.updateInterval = window.setInterval(() => {
-      this.updateTrajectories();
-    }, this.requestIntervalSeconds * 1000);
-  }
-
-  /**
-   * Stop the update of trajectories.
-   * @private
-   */
-  stopUpdateTrajectories() {
-    clearInterval(this.updateInterval);
-  }
-
-  /**
-   * Update the trajectories
-   * @private
-   */
-  updateTrajectories() {
-    this.abortFetchTrajectories();
-    this.abortController = new AbortController();
-    this.api
-      .fetchTrajectories(
-        this.getParams({
-          attr_det: 1,
-        }),
-        this.abortController,
-      )
-      .then((trajectories) => {
-        this.tracker.setTrajectories(trajectories);
-      });
-  }
-
-  /**
-   * Define the style of the vehicle.
-   *
-   * @param {Object} props Properties
-   * @private
-   */
-  defaultStyle(props) {
-    const { type, name, id, color, textColor, delay, cancelled } = props;
-    const z = Math.min(Math.floor(this.currentZoom || 1), 16);
-    const hover = this.tracker.hoverVehicleId === id;
-    const selected = this.selectedVehicleId === id;
-    const key = `${z}${type}${name}${delay}${hover}${selected}`;
-
-    if (!this.styleCache[key]) {
-      let radius = getRadius(type, z);
-
-      if (hover || selected) {
-        radius += 5;
-      }
-      const margin = 1;
-      const radiusDelay = radius + 2;
-      const origin = radiusDelay + margin;
-
-      const canvas = document.createElement('canvas');
-      canvas.width = radiusDelay * 2 + margin * 2 + 100;
-      canvas.height = radiusDelay * 2 + margin * 2;
-      const ctx = canvas.getContext('2d');
-
-      if (delay !== null) {
-        // Draw delay background
-        ctx.save();
-        ctx.beginPath();
-        ctx.arc(origin, origin, radiusDelay, 0, 2 * Math.PI, false);
-        ctx.fillStyle = getDelayColor(delay, cancelled);
-        ctx.filter = 'blur(1px)';
-        ctx.fill();
-        ctx.restore();
-      }
-
-      // Show delay if feature is hovered or if delay is above 5mins.
-      if (hover || delay >= this.delayDisplay) {
-        // Draw delay text
-        ctx.save();
-        ctx.textAlign = 'left';
-        ctx.textBaseline = 'middle';
-        ctx.font = `bold ${Math.max(
-          14,
-          Math.min(17, radius * 1.2),
-        )}px arial, sans-serif`;
-        ctx.fillStyle = getDelayColor(delay, cancelled);
-
-        ctx.strokeStyle = this.delayOutlineColor;
-        ctx.lineWidth = 1.5;
-        ctx.strokeText(getDelayText(delay, cancelled), origin * 2, origin);
-        ctx.fillText(getDelayText(delay, cancelled), origin * 2, origin);
-        ctx.restore();
-      }
-
-      ctx.beginPath();
-      ctx.arc(origin, origin, radius, 0, 2 * Math.PI, false);
-      if (!this.useDelayStyle) {
-        ctx.fillStyle = color || getBgColor(type);
-        ctx.fill();
-      } else {
-        ctx.fillStyle = getDelayColor(delay, cancelled);
-        ctx.fill();
-      }
-
-      ctx.lineWidth = 1;
-      ctx.strokeStyle = '#000000';
-      ctx.stroke();
-
-      const markerSize = radius * 2;
-      if (radius > 10) {
-        const shortname =
-          type === 'Rail' && name.length > 3 ? name.substring(0, 2) : name;
-        const fontSize = Math.max(radius, 10);
-        const textSize = getTextSize(ctx, markerSize, shortname, fontSize);
-
-        ctx.textBaseline = 'middle';
-        ctx.textAlign = 'center';
-        ctx.fillStyle = !this.useDelayStyle
-          ? textColor || getTextColor(type)
-          : '#000000';
-        ctx.font = `bold ${textSize}px Arial`;
-        ctx.fillText(shortname, origin, origin);
-      }
-      this.styleCache[key] = canvas;
-    }
-
-    return this.styleCache[key];
-  }
 }
-
-TrajservLayer.LINE_FILTER = 'publishedlinename';
-TrajservLayer.ROUTE_FILTER = 'tripnumber';
-TrajservLayer.OPERATOR_FILTER = 'operator';
 
 export default TrajservLayer;
